full rewrite of the grammar
A complete rewrite of the grammar, with almost no parts used from the
original.

The parser is now capable of parsing ~97% of all files within the
nim-lang/Nim repository, however the parser size is extremely large:
132MiB.

Some features are currently disabled to prevent the parser from
exceeding tree-sitter's own limit.