omit terminator for elif/else/except/finally
Previously, we omitted these terminators conditonally
(when the parser said it was expecting these). However
as the parser have to report states to the external scanner,
it blew up the parser size and state count exponentially.

At the cost of accuracy, these tokens are now unconditionally
treated as continuations and will not have terminator reported.

With this change the parser size has reduced back to 59MiB,
half of what it was before.