scanner: unconditional layout termination at pair ends
Since `)`, `]`, `}` and `.}` are paired, there is no need to
verify whether those tokens are being expected. Instead we
can rely solely on whether layout termination is expected.

To visualize this idea, we use `{` and `}` as layout_start and
end, with `;` being layout_terminator:

   if foo:
     {
     (foo)
     #   ^ a `)` is expected, and none of `;` or `}` is
     #     expected here.
     }

  if bar:
    {
    (if x:
      {
      y;})
    #  ^ as `{` was emitted, `;` and/or `}` is expected
    #    and `)` won't be expected.
    }

In the two possible cases, it is observed that whenever `)` is
expected, neither `;` nor `}` will be expected, and vice versa.
Thus there is no need to check the status of these closing tokens.

With this change, the parser is now 60MiB and only requires 7GiB of
RAM to generate.