Don't track contiguous status on novel delimiter edges
This is harder to reason about, and
2e6666041fa8b3c70077042b988843ca4f241026 did not include a motivating
test case.

Removing contiguous status is a minor perf improvement (2% reduction
in instructions), makes the code simpler, and does not significantly
affect diffing results.

Of the two sample files that have changed, the erlang_before.erl file
has improved and nest_before.rs is neutral.