scanner: rewrite in C
As the tree-sitter project actively discourage C++ scanners, it was
clear that we have to adapt our parser to C to maintain interop
with future direction of the ecosystem.

The logic and code structure of this rewrite is nearly identical to the
C++ version. This is intentional to prevent behavior bugs from creeping
in, and an overhaul will come at a later date.

Right now this code runs with no crashes and seems to work as well as
its C++ predecessor.

Fixes https://github.com/alaviss/tree-sitter-nim/issues/44