Remove support for 'try' from the parser
This was 'soft-removed' in the past by removing the highlights and test
cases but we held off on fully removing the parser support. This
finishes the job so that 'try' is no longer treated as a keyword.
This should fix the integration tests. Previously they would fail
against some code in the stdlib which used 'try' as a regular variable
name.