Simplify id.js, add infix expressions, fix operator declarations, rename list to array, remove some Haskell legacy, align type literals with PS spec
- Add infix expressions. Previously there were only unqualified infix identifiers, but PS allows placing any expression in infix form.
Highlights queries act inconsistently across different renderers. The default choice is to highlight everything as an operator, but it only works for `tree-sitter highlight`. In Helix, I could get specific child nodes to be highlighted, but it's limited so the choice is to only highlight identifiers and qualified identifiers as operators, and delegate the rest to respective nodes. In Kakoune I couldn't get anything to work. I don't have a Vim/Emacs setup to test it further. Leaving it as it is because it's satisfactory.

- Remove infix identifiers from function patterns
This is not a thing in PS.

- Remove implicit parameters
Not a thing in PS. Later they will be replaced with type holes.

- Remove `fixity` declaration and introduce `operator_declaration`.
They have different syntax.

- Rename lists to arrays

- Remove `label` expressions/identifiers
Not a thing in PS.

- Simplify `id.js`
Removed some Haskell-specific things and added a few comments to make this module easier to understand.

- Remove type-level literals for lists and tuples
Not a thing in PS.

- Add `hiding` keyword to highlight queries