Try fix `in` in ado blocks
A complication here is that do/ado blocks are allowed to use `let` binders,
which in turn has ambiguous interpretation wrt `let in` expressions, so something like
```
ado
  let a = 1
  in a
```
becomes rather tricky. The solution is to remove `let in` expressions
from the list of options for a do/ado-block binder-indentation level statements.
They aren't allowed in PS itself, unless parenthesized. But this is buried
a bit deep inside a hierarchy of parsers so the code involves some duplication,
basically writing out almost the same thing twice, for do/ado blocks and for normal
notation. It still doesn't handle cases where `in` terminator is one the same line
with a `let` binder, so it will require some more fiddling.