Add --hidden and --no-ignore options to walk all files in directory
The ignore::Walk excludes various files by default. It might makes sense if
user runs difft command manually, but difft can also be used as a directory
diff backend of version control tools. In that case, extra filtering is rather
problematic than useful.

The --hidden and --no-ignore flags are copied from ripgrep. We can add more
--no-ignore-<kind> as needed. For my use case, these two flags are good enough.
Alternatively, we can change the default to not exclude hidden (or dot) files.
That's sufficient for my needs, but I'm not sure if that's desired. The major
problem I have right now is that .gitignore changes are excluded (because it's
a "hidden" file.)

https://github.com/BurntSushi/ripgrep/blob/14.0.3/crates/core/flags/hiargs.rs#L889
https://github.com/BurntSushi/ripgrep/blob/14.0.3/crates/ignore/src/dir.rs#L586