scanner: allow end on EOF and use column for flag invalidation
Allowing layout_end on EOF allows the parser to close and isolate
grammar portions where layout_end is typically not allowed (ie. in
parentheses), which enabled better error recovery.

The empty termination hack does not seem to contribute anymore with this
change and was removed.

With this change, we also remove the use of synchronize node for fixing
after newline. Instead it is cleared based on column position at the
start of the lexer. This should reduce errors from node reuse, but is
more or less a hunch since it is hard to test incremental parsing.