fix: different approach to fix off-by-one without affecting display
This approach removes the additional empty line that would cause
the display logic to disagree in line count.

The major difference is that this new approach does not change the
output of the display implementation, therefore, no test files need to
be updated, and no additional line is printed to the end-user.

As for the cause of the index out-of-bounds, this is simply a
disagreement between line sources, such as the `line-numbers`
crate and Tree Sitter, which all reports the correct number of lines,
with the `str::lines` iterator which does not iterate over the last
new line character if the last line is empty.