feat: specify real path of file in environment
The resolution of #620 does not help for the case of git invocations comparing
two git commits, like `git dft main...` or `git dft branch1...branch2`, because
in those cases both the lhs and the rhs path are temporary files, obscuring the
real path to the file within git.

In those cases, git makes the real path available as the MERGED shell variable,
for the purposes of constructing a difftool command line in git configuration.

So in Difftastic, let the real path be specified by an environment variable,
and suggest populating that with "$MERGED" in the git difftool configuration.