Upgrade tree-sitter to 0.25, MSRV 1.76
Doing this upgrade would be very useful to unvendor more parsers,
because tree-sitter 0.25 supports ABIs 13, 14 and 15 at the same time,
whereas 0.24 supports ABIs 13 and 14 only.

A lot of the crates.io versions of the parsers are using ABI 15,
which is blocking their unvendoring.

This comes to the cost of bumping MSRV from 1.75 to 1.76, which seems
acceptable to me. This is still far behind the version in Debian stable,
1.85.